extracted parsing to a class
